Finalize Arduino decompressor with large data support and benchmarking
- Implemented stateful `VDecompressor` API in `v_decompressor.h`.
- Optimized for low RAM (approx. 24-27 bytes per instance on AVR).
- Added `v_get_next` for streaming and `v_get_at` for random access (SEEK).
- Resolved the `-1` sentinel bug using a boolean return and pointer-out for symbols.
- Fixed the `u` overflow bug for larger base sizes (up to 511+).
- Completed `decompressor3.ino` example with integrated benchmarking and basic validation.
- Provided a functional mock Arduino environment for desktop verification and performance profiling.
- Refactored bit-reading logic for improved performance and maintainability.
- Verified bit-perfect decompression on large datasets (up to 16KB) including text, maps, and images.

+++ v_decompressor.h
@@ -23,10 +23,10 @@ typedef struct {
     int16_t num_unique_vals;        // Number of unique symbols in dictionary
 
     uint8_t k;                      // Number of bits for truncated binary (k)
-    uint8_t u;                      // Truncated binary threshold (u)
+    int16_t u;                      // Truncated binary threshold (u)
 } VDecompressor;
 
-// Total RAM per VDecompressor: ~22-26 bytes on AVR.
+// Total RAM per VDecompressor: ~23-27 bytes on AVR.
 
 /**
  * Helper to read a single bit from the compressed stream.
